refactor post types
