+++ app/components/building-blocks/post-info-bar.tsx
@@ -1,7 +1,8 @@
 import moment from 'moment'
 import { useState } from 'react'
 import { type FallacyList } from '#app/modules/fallacies/fallacy-types.ts'
-import { type PostState, type Post } from '#app/types/api-types.ts'
+import { type Post } from '#app/modules/posts/post-types.ts'
+import { type PostState } from '#app/types/api-types.ts'
 
 export function PostInfoBar({
 	post,
