+++ app/routes/_actions+/create-poll.tsx
@@ -2,7 +2,7 @@ import { type ActionFunctionArgs } from '@remix-run/node'
 import { z } from 'zod'
 import { db } from '#app/db.ts'
 import { getOrCreatePoll } from '#app/modules/posts/polls/poll-repository.ts'
-import { PollType } from '#app/types/api-types.ts'
+import { PollType } from '#app/modules/posts/post-types.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
 
 const PollCreationDtoSchema = z.object({
