+++ app/routes/polls.tsx
@@ -6,12 +6,9 @@ import { PostContent } from '#app/components/building-blocks/post-content.tsx'
 import { Markdown } from '#app/components/markdown.tsx'
 import { Icon } from '#app/components/ui/icon.tsx'
 import { db } from '#app/db.ts'
+import { type PollPagePost } from '#app/modules/posts/post-types.ts'
 import { getChronologicalPolls } from '#app/modules/scoring/ranking-service.ts'
-import {
-	type Artefact,
-	type Quote,
-	type PollPagePost,
-} from '#app/types/api-types.ts'
+import { type Artefact, type Quote } from '#app/types/api-types.ts'
 
 export async function loader() {
 	const feed = await db.transaction().execute(async trx => {
