+++ app/routes/stats.$postId.tsx
@@ -6,9 +6,9 @@ import { GeneralErrorBoundary } from '#app/components/error-boundary.tsx'
 import { Markdown } from '#app/components/markdown.tsx'
 import { db } from '#app/db.ts'
 import { getStatsPost } from '#app/modules/posts/post-repository.ts'
+import { type StatsPost } from '#app/modules/posts/post-types.ts'
 import { getEffects } from '#app/modules/scoring/effect-repository.ts'
 import { relativeEntropy } from '#app/modules/scoring/scoring-utils.ts'
-import { type StatsPost } from '#app/types/api-types.ts'
 import { type DBEffect } from '#app/types/db-types.ts'
 
 const postIdSchema = z.coerce.number()
